implemented main()
So this is a new pkglist-query implementation which uses zpkglist
and runs two threads: the first thread reads the header blobs, and
the second one does the rest of the job.  But the first thread can also
"help" the second thread.  In other words, the first thread balances
between loading more blobs and coping with the already loaded ones.
The whole process is seamless: the queue discipline ensures that
the results are output in the original order.

This works very well with simple queries like '%{Name}\n'.
(Heavy queries like '[%{FileNames}\n]' is another story.)

$ set /ALT/Sisyphus/x86_64/base/pkglist.classic.zst
$ time ./pkglist-query '%{Name}\n' $1 >/dev/null
0.23s user 0.01s system 127% cpu 0.186 total

This is even faster than piping from zstd to the earlier
pkglist-unmets implementation.

$ time (zstd -d <$1 |pkglist-query '%{Name}\n' - ) >/dev/null
0.25s user 0.05s system 133% cpu 0.226 total