set default NQ to 128, optimize linear searches
The problem with heavy queries like [%{FileNames}\n] is that they
are very susceptible to the number of files in a package.  When
a really big header in a bloated pkglist is followed by a bunch of
small headers, the queue gets stalled after all the small headers
are queried, but the big header is still being processed.