short-lived headers, use cookies to identify entries
Before this change, the blob pointer was used to identify entries.
This increases the number of malloc chunks beyond the necessity and
hampers parallelism, because the headers are freed under the lock.
A much better scenario is to free the header along with the blob ASAP.
Entries can instead be identified with a sequence number, which I call
cookies, because they're specific to the cooking stage.