the main thread can help even before the queue is full
I noticed that performance degrades slightly when the queue
is big, the query is simple, and the headers are non-bloated.
This is likely due to L1/L2 cache effects: the blobs have to
loaded from memory all over again.

So the queue should generally be kept rather small.  The only
justifiable case for the queue to grow to its full is when the
worker thread is doing a heavy job, and the main thread keeps
pushing new blobs to the queue, because the only alternative
for the main thread is to stall.

This new needMoreAid() strategy has been carefully profiled.